Move auth-related views and serializers to auth subproject
