+++ app/data/views.py
@@ -11,16 +11,9 @@ from django.db.models import (Case,
 
 from django_redis import get_redis_connection
 
-from django.contrib.auth.models import User, Group
-from rest_framework import permissions, status, viewsets
-from rest_framework.authtoken.models import Token
-from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.decorators import list_route
 from rest_framework.response import Response
 from rest_framework.exceptions import ParseError
-
-from data.serializers import UserSerializer, GroupSerializer
-from driver_auth.permissions import IsAdminOrReadSelfOnly, IsAdminOrReadOnly
 
 from ashlar import views
 from data import transformers
@@ -133,43 +126,3 @@ class DriverRecordViewSet(views.RecordViewSet):
         # to store the data exactly as it is.
         redis_conn = get_redis_connection('default')
         redis_conn.set(token, sql.encode('utf-8'))
-
-
-class UserViewSet(viewsets.ModelViewSet):
-    """
-    API endpoint that allows users to be viewed or edited.
-    """
-
-    serializer_class = UserSerializer
-    permission_classes = (IsAdminOrReadSelfOnly,)
-    queryset = User.objects.all().order_by('-date_joined')
-
-    def get_queryset(self):
-        """Limit non-admin users to only see their own info"""
-        user = self.request.user
-        if user.is_staff:
-            return self.queryset
-        else:
-            return self.queryset.filter(id=user.id)
-
-
-class GroupViewSet(viewsets.ModelViewSet):
-    """
-    API endpoint that allows groups to be viewed or edited.
-    """
-    queryset = Group.objects.all()
-    serializer_class = GroupSerializer
-    permission_classes = (IsAdminOrReadOnly,)
-
-
-class DriverObtainAuthToken(ObtainAuthToken):
-    def post(self, request):
-        serializer = self.serializer_class(data=request.DATA)
-        if serializer.is_valid(raise_exception=True):
-            user = serializer.validated_data['user']
-            token, created = Token.objects.get_or_create(user=user)
-            return Response({'token': token.key, 'user': token.user_id})
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-obtain_auth_token = DriverObtainAuthToken.as_view()
