+++ app/driver/urls.py
@@ -7,6 +7,7 @@ from rest_framework import routers
 from ashlar import views as a_views
 
 from data import views as data_views
+from driver_auth import views as auth_views
 from user_filters import views as filt_views
 
 router = routers.DefaultRouter()
@@ -18,14 +19,14 @@ router.register('recordschemas', a_views.RecordSchemaViewSet)
 router.register('recordtypes', a_views.RecordTypeViewSet)
 
 # user management
-router.register(r'users', data_views.UserViewSet)
-router.register(r'groups', data_views.GroupViewSet)
+router.register(r'users', auth_views.UserViewSet)
+router.register(r'groups', auth_views.GroupViewSet)
 
 urlpatterns = [
     url(r'^admin/', include(admin.site.urls)),
     url(r'^api/', include(router.urls)),
     # get token for given username/password
-    url(r'^api-token-auth/', data_views.obtain_auth_token),
+    url(r'^api-token-auth/', auth_views.obtain_auth_token),
     # OIDC
     url(r'openid/', include('djangooidc.urls')),
 ]
