+++ app/driver_auth/views.py
@@ -1,0 +1,46 @@
+from django.contrib.auth.models import User, Group
+from rest_framework import permissions, status, viewsets
+from rest_framework.authtoken.models import Token
+from rest_framework.authtoken.views import ObtainAuthToken
+
+from driver_auth.serializers import UserSerializer, GroupSerializer
+from driver_auth.permissions import IsAdminOrReadSelfOnly, IsAdminOrReadOnly
+
+class UserViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows users to be viewed or edited.
+    """
+
+    serializer_class = UserSerializer
+    permission_classes = (IsAdminOrReadSelfOnly,)
+    queryset = User.objects.all().order_by('-date_joined')
+
+    def get_queryset(self):
+        """Limit non-admin users to only see their own info"""
+        user = self.request.user
+        if user.is_staff:
+            return self.queryset
+        else:
+            return self.queryset.filter(id=user.id)
+
+
+class GroupViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows groups to be viewed or edited.
+    """
+    queryset = Group.objects.all()
+    serializer_class = GroupSerializer
+    permission_classes = (IsAdminOrReadOnly,)
+
+
+class DriverObtainAuthToken(ObtainAuthToken):
+    def post(self, request):
+        serializer = self.serializer_class(data=request.DATA)
+        if serializer.is_valid(raise_exception=True):
+            user = serializer.validated_data['user']
+            token, created = Token.objects.get_or_create(user=user)
+            return Response({'token': token.key, 'user': token.user_id})
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+obtain_auth_token = DriverObtainAuthToken.as_view()
